# Conflicts:
#	virtual_machine/bonus/check_if_refresh.c
#	virtual_machine/bonus/ft_kbhit.c
#	virtual_machine/bonus/put_caret_on_map.c
#	virtual_machine/bonus/time.c
#	virtual_machine/bonus/visualisate.c
#	virtual_machine/bonus/visualisation.h
#	virtual_machine/bonus/w_print_usage.c
#	virtual_machine/get_args_and_values_part_2.c
#	virtual_machine/vm.h

+++ virtual_machine/lfork_operation.c
@@ -17,7 +17,7 @@ void    do_lfork_func(t_vm *main_struct, t_car *car)
 
 	i = 0;
 	arg_size = car->op_tabble.codage_octal == 0 ? DIR_SIZE : IND_SIZE;
-	dist = (int) get_short_from_byte_code(main_struct->map + car->pos + 1, car->op_tabble
+	dist = get_short_from_byte_code(main_struct->map + car->pos + 1, car->op_tabble
 			.codage_octal);
 	while (i < main_struct->players_nbr)
 	{
